Update Starter to newer Key code

+++ Code/Starter/FaceRecognition/Files/CustomFaceRecognizer.mm
@@ -147,7 +147,7 @@
 
     if (sqlite3_prepare_v2(_db, insertSQL, -1, &statement, nil) == SQLITE_OK) {
         sqlite3_bind_int(statement, 1, personID);
-        sqlite3_bind_blob(statement, 2, serialized.bytes, serialized.length, SQLITE_TRANSIENT);
+        sqlite3_bind_blob(statement, 2, serialized.bytes, (int) serialized.length, SQLITE_TRANSIENT);
         sqlite3_step(statement);
     }
